+++ packages/extended-text-entry/configure/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.6](https://github.com/pie-framework/pie-elements/compare/@pie-element/extended-text-entry-configure@3.0.5...@pie-element/extended-text-entry-configure@3.0.6) (2019-06-17)
+
+**Note:** Version bump only for package @pie-element/extended-text-entry-configure
+
+
+
+
+
+## [3.0.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/extended-text-entry-configure@3.0.4...@pie-element/extended-text-entry-configure@3.0.5) (2019-06-12)
+
+**Note:** Version bump only for package @pie-element/extended-text-entry-configure
+
+
+
+
+
+## [3.0.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/extended-text-entry-configure@3.0.3...@pie-element/extended-text-entry-configure@3.0.4) (2019-06-07)
+
+**Note:** Version bump only for package @pie-element/extended-text-entry-configure
+
+
+
+
+
+## [3.0.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/extended-text-entry-configure@3.0.2...@pie-element/extended-text-entry-configure@3.0.3) (2019-06-06)
+
+**Note:** Version bump only for package @pie-element/extended-text-entry-configure
+
+
+
+
+
+## [3.0.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/extended-text-entry-configure@3.0.1...@pie-element/extended-text-entry-configure@3.0.2) (2019-06-05)
+
+**Note:** Version bump only for package @pie-element/extended-text-entry-configure
+
+
+
+
+
 ## [3.0.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/extended-text-entry-configure@3.0.0...@pie-element/extended-text-entry-configure@3.0.1) (2019-05-08)
 
 **Note:** Version bump only for package @pie-element/extended-text-entry-configure