+++ packages/inline-choice/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.17](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-choice@2.5.16...@pie-element/inline-choice@2.5.17) (2019-06-17)
+
+**Note:** Version bump only for package @pie-element/inline-choice
+
+
+
+
+
+## [2.5.16](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-choice@2.5.15...@pie-element/inline-choice@2.5.16) (2019-06-12)
+
+**Note:** Version bump only for package @pie-element/inline-choice
+
+
+
+
+
+## [2.5.15](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-choice@2.5.14...@pie-element/inline-choice@2.5.15) (2019-06-07)
+
+**Note:** Version bump only for package @pie-element/inline-choice
+
+
+
+
+
+## [2.5.14](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-choice@2.5.13...@pie-element/inline-choice@2.5.14) (2019-06-06)
+
+**Note:** Version bump only for package @pie-element/inline-choice
+
+
+
+
+
+## [2.5.13](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-choice@2.5.12...@pie-element/inline-choice@2.5.13) (2019-06-05)
+
+**Note:** Version bump only for package @pie-element/inline-choice
+
+
+
+
+
+## [2.5.12](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-choice@2.5.11...@pie-element/inline-choice@2.5.12) (2019-05-23)
+
+**Note:** Version bump only for package @pie-element/inline-choice
+
+
+
+
+
 ## [2.5.11](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-choice@2.5.10...@pie-element/inline-choice@2.5.11) (2019-05-21)
 
 **Note:** Version bump only for package @pie-element/inline-choice