+++ packages/placement-ordering/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.1.7](https://github.com/pie-framework/pie-elements/compare/@pie-element/placement-ordering@5.1.6...@pie-element/placement-ordering@5.1.7) (2019-06-17)
+
+**Note:** Version bump only for package @pie-element/placement-ordering
+
+
+
+
+
+## [5.1.6](https://github.com/pie-framework/pie-elements/compare/@pie-element/placement-ordering@5.1.5...@pie-element/placement-ordering@5.1.6) (2019-06-12)
+
+**Note:** Version bump only for package @pie-element/placement-ordering
+
+
+
+
+
+## [5.1.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/placement-ordering@5.1.4...@pie-element/placement-ordering@5.1.5) (2019-06-07)
+
+**Note:** Version bump only for package @pie-element/placement-ordering
+
+
+
+
+
+## [5.1.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/placement-ordering@5.1.3...@pie-element/placement-ordering@5.1.4) (2019-06-06)
+
+**Note:** Version bump only for package @pie-element/placement-ordering
+
+
+
+
+
+## [5.1.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/placement-ordering@5.1.2...@pie-element/placement-ordering@5.1.3) (2019-06-05)
+
+**Note:** Version bump only for package @pie-element/placement-ordering
+
+
+
+
+
+## [5.1.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/placement-ordering@5.1.1...@pie-element/placement-ordering@5.1.2) (2019-05-23)
+
+**Note:** Version bump only for package @pie-element/placement-ordering
+
+
+
+
+
 ## [5.1.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/placement-ordering@5.1.0...@pie-element/placement-ordering@5.1.1) (2019-05-21)